requirements.txt
asttokens==3.0.0
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
Flask==3.1.2
flask-cors==6.0.1
Flask-MySQLdb==2.0.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==9.4.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
python-dateutil==2.9.0.post0
pywin32==310
pyzmq==26.4.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.13
Werkzeug==3.1.3
